[TM4C123_ACOMP] Update Common utils in MCU library

+++ TM4C123_ACOMP/xDriver_MCU/Common/xSource/MCU_CheckParams.c
@@ -23,12 +23,11 @@
  */
 #include <xDriver_MCU/Common/xHeader/MCU_CheckParams.h>
 
-uint32_t MCU__u32CheckPatams(uint32_t u32Module,uint32_t u32ModuleMax)
+uint32_t MCU__u32CheckPatams(uint32_t u32Module, uint32_t u32ModuleMax)
 {
-    if((uint32_t)u32ModuleMax<u32Module)
+    if(u32ModuleMax < u32Module)
     {
-        u32Module=(uint32_t)u32ModuleMax;
+        u32Module = u32ModuleMax;
     }
     return u32Module;
 }
-
